ref: changing checkIfUserHasPlant-function to a common function
