+++ src/common/plants/common.ts
@@ -1,4 +1,5 @@
 import { queryGetPlantByPlantName } from '../../database/plants/queryGetPlantByPlantName';
+import { queryGetPlantByPlantIdAndUserId } from '../../database/plants/queryGetPlantByPlantIdAndUserId';
 import { NotFoundError } from '../../errors/error';
 import { errorMessages } from '../../errors/errorMessages';
 import { Plant } from '../../types/plant/plant';
@@ -19,4 +20,18 @@ export const doesPlantExist = (plant: Plant): void => {
 export const doesPlantLocationExist = (plantLocation: PlantLocation): void => {
 	if (isNullOrUndefined(plantLocation))
 		throw new NotFoundError(errorMessages.plantLocationNotFound);
+};
+
+export const checkIfUserHasPlant = async (
+	plantId: string,
+	userId: string,
+): Promise<void> => {
+	const plant: Plant = (await queryGetPlantByPlantIdAndUserId(
+		plantId,
+		userId,
+	)) as Plant;
+
+	if (isNullOrUndefined(plant)) {
+		throw new NotFoundError(errorMessages.plantNotFound);
+	}
 };
